+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-29T13:09:23.492Z",
+    "2026-05-30T09:13:59.913Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {